requirements.txt
tweepy
google-genai
fastmcp